refactor(firebase): split the service by concern, one module per collection
`services/firebase.ts` was 23 exports over 9 unrelated concerns — auth, chat,
boards, schedule, presence, custom actions, a data wipe, and the SDK handles —
with no two concerns sharing a caller. Every consumer imported the union, and
the single test mock stubbed all of it (plus syncService).

Now `services/firebase/`:

- `app.ts` — the only module that initializes Firebase and hands out handles, so
  there is no second route to the SDK and no second app instance.
- `auth.ts` — sign-in, registration, anonymous upgrade, sign-out, data wipe.
- `chat.ts` — a room's messages, both directions: send guard, delete, and the
  read window (3h, newest 50).
- `boards.ts` — shared boards by content hash; returns domain values only.
- `schedule.ts` — the upcoming list plus its writers, owning the Timestamp ↔
  Dayjs representation in both directions.
- `customActions.ts` — the community-suggestion write.

All 26 importers now name the concern they use; there is no barrel to re-widen
the interface. `lazyFirebase`'s dynamic chunk narrows from the whole service to
`auth`, which is all `context/auth.tsx` ever read off it.

The monolithic `src/__mocks__/firebase.ts` is replaced by one manual mock per
concern under `services/firebase/__mocks__/`, so a test that needs `sendMessage`
no longer gets a stubbed auth surface — or a stubbed syncService — for free.

No behaviour change: the bodies moved verbatim (the defect fixes shipped in
#1139), and the two `limit(50)` constants are now named for what they bound.
