+++ docs/engineering/data-and-sync.md
@@ -29,7 +29,7 @@ A **sync middleware** (`src/services/syncMiddleware.ts`) wraps Dexie writes: aft
 
 ## Firebase products & paths
 
-Initialized in `src/services/firebase.ts`.
+Initialized in `src/services/firebase/app.ts`, the only module that touches the SDK's app/handle APIs; every other module takes `db` from it.
 
 - **Auth** — anonymous, email/password, and Google. Anonymous accounts can be upgraded in place (`linkWithCredential`) keeping the same UID.
 - **Firestore** — with **offline persistence** (`persistentLocalCache` + `persistentMultipleTabManager`), falling back to in-memory cache if IndexedDB is unavailable (e.g. private browsing). Collections:
@@ -155,7 +155,7 @@ Includes user-created groups and tiles, and optionally your disabled-default lis
 
 Two share-by-link flows exist, both consumed from the Room view via URL query params:
 
-- **Game boards** (`?importBoard=<id>`) — when a board is activated/published, it is stored in `game-boards/{id}` (`getOrCreateBoard` in `firebase.ts`, SHA-256 dedup, 30-day TTL; returns `{ id }` — the only thing callers use, and previously two different Firestore classes under one declared type) and a chat message carries the id. Opening a link with `?importBoard=<id>` is handled by `src/hooks/useUrlImport.ts`: it fetches the board, parses its `gameBoard`/`settings`, and upserts it locally as the active board. Share URLs are built in `MessageList/Message` and `gameSettingsMessage.ts`.
+- **Game boards** (`?importBoard=<id>`) — when a board is activated/published, it is stored in `game-boards/{id}` (`getOrCreateBoard` in `firebase/boards.ts`, SHA-256 dedup, 30-day TTL; returns `{ id }` — the only thing callers use, and previously two different Firestore classes under one declared type) and a chat message carries the id. Opening a link with `?importBoard=<id>` is handled by `src/hooks/useUrlImport.ts`: it fetches the board, parses its `gameBoard`/`settings`, and upserts it locally as the active board. Share URLs are built in `MessageList/Message` and `gameSettingsMessage.ts`.
 - **Content packs** (`?importPack=<id>`) — durable bundles scoped to one or more author-selected custom groups (their custom tiles + group defs), published to `content-packs/{id}` (`src/services/contentPacks.ts`, reusing the export/import serialization). `src/hooks/useUrlPackImport.ts` fetches the pack and opens a **full-dump preview dialog** (`CustomTileDialog/Packs/PackImportDialog`) listing every group + action card before the user confirms a single **Import**. Import is **copy-only**: `importData` clones the contents into the user's own custom content, stamping a lightweight `packId` + `packName` for attribution and re-import dedupe (`ViewCustomTiles` shows a "From {pack}" chip). There is no subscription, version propagation, or auto-update. Publishing (multi-group select + visibility) and import-by-code live in the **Packs** accordion (`ctPacks`); expanding it swaps the dialog's right pane to the public directory (`PackDirectory`).
 
 **Visibility + directory:** each pack is `public` (directory-listed) or `private` (unlisted, importable by code). `listPublicPacks` queries `content-packs` by `where(visibility=public, gameMode, locale)` + `orderBy(createdAt desc)` with cursor pagination (composite index in `firestore.indexes.json`); name/tag filtering is client-side over the loaded page. Denormalized `tileCount`/`groupCount`/`groupLabels` on the doc let directory cards render without parsing `contents`. Publishing **public** requires a permanent (non-anonymous) account; anonymous users publish private only. Moderation is report → `reports/` → manual Firebase-Console takedown (`admin` custom claim); no in-app admin UI. An author may **republish** (bumps `packVersion`, refreshes the listing), but existing importers are never notified — their copy is unaffected.
