+++ src/components/MessageInput/index.tsx
@@ -2,7 +2,7 @@ import { useState, FormEvent, ChangeEvent, KeyboardEvent } from 'react';
 import { IconButton, InputAdornment, TextField } from '@mui/material';
 import SendIcon from '@mui/icons-material/Send';
 import useAuth from '@/context/hooks/useAuth';
-import { sendMessage } from '@/services/firebase';
+import { sendMessage } from '@/services/firebase/chat';
 import './styles.css';
 import { Trans, useTranslation } from 'react-i18next';
 import { Camera, CameraResultType, Photo as CameraPhoto } from '@capacitor/camera';
