+++ src/components/auth/Login.tsx
@@ -1,7 +1,7 @@
 import { useState, FormEvent, ChangeEvent } from 'react';
 import { Box, Button, TextField, Typography, Alert, CircularProgress } from '@mui/material';
 import { Trans, useTranslation } from 'react-i18next';
-import { loginWithEmail } from '@/services/firebase';
+import { loginWithEmail } from '@/services/firebase/auth';
 
 interface LoginProps {
   onSuccess?: () => void;
