+++ src/hooks/usePlayerMove.ts
@@ -1,6 +1,6 @@
 import { useCallback, useEffect, useRef, useState } from 'react';
 import { useTranslation } from 'react-i18next';
-import { sendMessage } from '@/services/firebase';
+import { sendMessage } from '@/services/firebase/chat';
 import useAuth from '@/context/hooks/useAuth';
 import actionStringReplacement from '@/services/actionStringReplacement';
 import usePlayerList from './usePlayerList';
