+++ src/hooks/useReturnToStart.ts
@@ -1,6 +1,6 @@
 import { useTranslation } from 'react-i18next';
 import { Params, useParams } from 'react-router-dom';
-import { sendMessage } from '@/services/firebase';
+import { sendMessage } from '@/services/firebase/chat';
 import { useCallback } from 'react';
 import useAuth from '@/context/hooks/useAuth';
 import { TurnFields } from '@/types/Message';
