+++ src/hooks/useUrlImport.ts
@@ -1,4 +1,4 @@
-import { getBoard } from '@/services/firebase';
+import { getBoard } from '@/services/firebase/boards';
 import { useEffect, useState, useCallback } from 'react';
 import { useTranslation } from 'react-i18next';
 import { useSearchParams } from 'react-router-dom';
