+++ src/services/firebase/__mocks__/app.ts
@@ -1,0 +1,4 @@
+import { vi } from 'vitest';
+
+export const db = {};
+export const getRealtimeDb = vi.fn(() => ({}));
