+++ src/services/firebase/__mocks__/boards.ts
@@ -1,0 +1,4 @@
+import { vi } from 'vitest';
+
+export const getOrCreateBoard = vi.fn().mockResolvedValue({ id: 'mock-board-id' });
+export const getBoard = vi.fn().mockResolvedValue(undefined);
