+++ src/services/firebase/__mocks__/chat.ts
@@ -1,0 +1,5 @@
+import { vi } from 'vitest';
+
+export const getMessages = vi.fn(() => vi.fn());
+export const sendMessage = vi.fn().mockResolvedValue(undefined);
+export const deleteMessage = vi.fn().mockResolvedValue(undefined);
