+++ src/services/firebase/__mocks__/customActions.ts
@@ -1,0 +1,3 @@
+import { vi } from 'vitest';
+
+export const submitCustomAction = vi.fn().mockResolvedValue(undefined);
