+++ src/services/roomSettingsService.ts
@@ -2,7 +2,7 @@ import { Settings } from '@/types/Settings';
 import { User } from '@/types';
 import { getSiteName } from '@/helpers/urls';
 import i18next from 'i18next';
-import { sendMessage } from '@/services/firebase';
+import { sendMessage } from '@/services/firebase/chat';
 
 function getRoomSettingsMessage(settings: Partial<Settings>): string {
   const { t } = i18next;
