+++ src/views/ManageGameBoards/index.tsx
@@ -25,7 +25,8 @@ import { t } from 'i18next';
 import { useSettings } from '@/stores/settingsStore';
 import { DEFAULT_TILE_COUNT } from '@/constants/boardConstants';
 import useAuth from '@/context/hooks/useAuth';
-import { getOrCreateBoard, sendMessage } from '@/services/firebase';
+import { getOrCreateBoard } from '@/services/firebase/boards';
+import { sendMessage } from '@/services/firebase/chat';
 import { isPublicRoom } from '@/helpers/strings';
 import { AlertState } from '@/types';
